Extract preferences out of config
